Initial GN checkin
work in progress, DON'T MERGE

android-emu-base library builds OK now

+++ android/android-emu/android/base/files/DecompressingStream.h
@@ -31,7 +31,7 @@ public:
 private:
     void* mLzStream;
     SmallFixedVector<char, 512> mBuffer;
-    int mBufferPos = 0;
+    size_t mBufferPos = 0;
 };
 
 }  // namespace base
